translations, renaming files, linted
